+++ src/net/session.h
@@ -30,6 +30,7 @@ using net::ip::tcp;
 namespace residue {
 class RequestHandler;
 class Registry;
+class Client;
 
 ///
 /// \brief Session object with each connection
@@ -69,6 +70,21 @@ public:
     ///
     void write(const std::string& s);
 
+    inline const std::string& id() const
+    {
+        return m_id;
+    }
+
+    inline void setClient(Client* client)
+    {
+        m_client = client;
+    }
+
+    inline const Client* client() const
+    {
+        return m_client;
+    }
+
     ///
     /// \brief Returns socket by const reference
     ///
@@ -87,6 +103,15 @@ public:
         return m_bytesSent;
     }
 private:
+    std::string m_id;
+    tcp::socket m_socket;
+    RequestHandler* m_requestHandler;
+    Client* m_client;
+    std::string m_name;
+    net::streambuf m_streamBuffer;
+
+    std::string m_bytesSent;
+    std::string m_bytesReceived;
 
     ///
     /// \brief Read incoming data and calls sendToHandler on the packet
@@ -107,14 +132,6 @@ private:
     /// \brief Send the packet bytes to the handler
     ///
     void sendToHandler(std::string&& incomingBytes);
-
-    tcp::socket m_socket;
-    RequestHandler* m_requestHandler;
-    std::string m_name;
-    net::streambuf m_streamBuffer;
-
-    std::string m_bytesSent;
-    std::string m_bytesReceived;
 };
 }