+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [Unreleased]
+### Fixes
+- Output for `sess --stats` fixed for multiple sessions
+
 ## [1.4.0]
 ### Updates
 - Based on standalone ASIO instead of boost